+++ app/app.js
@@ -9,6 +9,7 @@ window.run = function() {
         'config/baseConfig',
         'config/headerConfig',
         'plugins/anchoredLayoutPlugin',
+        'plugins/mobifyPreviewPlugin',
         'controllers/counterBadgeController',
         'app-components/deepLinkingServices',
         'app-controllers/tabBarController',
@@ -25,6 +26,7 @@ window.run = function() {
         BaseConfig,
         HeaderConfig,
         AnchoredLayoutPlugin,
+        MobifyPreviewPlugin,
         CounterBadgeController,
         DeepLinkingServices,
         TabBarController,
@@ -35,7 +37,6 @@ window.run = function() {
     ) {
         var deepLinkingServices = null;
         var errorControllerPromise = ErrorController.init();
-        var welcomeModalControllerPromise = WelcomeModalController.init(errorControllerPromise);
         var cartModalControllerPromise = CartModalController.init(errorControllerPromise);
         var cartEventHandlerPromise = cartModalControllerPromise.then(
             function(cartModalController) {
@@ -51,30 +52,6 @@ window.run = function() {
             counterBadgeController.updateCounterValue(3);
             return counterBadgeController;
         });
-
-        // Register RPC to expose whether it is possible to go back.
-        // This is necessary to determine whether to show/hide
-        // the back button on the error modal
-        var registerCanGoBackRpc = function(controller) {
-            Promise.join(welcomeModalControllerPromise, cartModalControllerPromise,
-            function(welcomeModal, cartModal) {
-                Astro.registerRpcMethod(AppRpc.names.appCanGoBack, [], function(res) {
-                    // We want to return true when the cart is showing since
-                    // calling `back` is equivalent to dismissing the cart
-                    if (cartModal.isShowing) {
-                        res.send(null, true);
-                    } else if (welcomeModal.isShowing) {
-                        welcomeModal.canGoBack().then(function(canGoBack) {
-                            res.send(null, canGoBack);
-                        });
-                    } else {
-                        controller.canGoBack().then(function(canGoBack) {
-                            res.send(null, canGoBack);
-                        });
-                    }
-                });
-            });
-        };
 
         var createTabBarLayout = function() {
             var layoutPromise = AnchoredLayoutPlugin.init();
@@ -102,8 +79,6 @@ window.run = function() {
                 tabBarControllerPromise,
                 layoutSetupPromise,
             function(tabBarController) {
-                registerCanGoBackRpc(tabBarController);
-
                 return tabBarController;
             });
         };
@@ -131,36 +106,52 @@ window.run = function() {
                     drawerController.navigateToNewRootView(url, title);
                     res.send(null, 'success');
                 });
-                registerCanGoBackRpc(drawerController);
-
                 return drawerController;
             });
         };
 
-        var appLayoutPromise = Application.getOSInformation().then(function(osInfo) {
-            if (BaseConfig.useTabLayout) {
-                return createTabBarLayout(counterBadgeControllerPromise);
-            } else {
-                return createDrawerLayout(counterBadgeControllerPromise);
-            }
-        });
+        var createLayout = function() {
+            return BaseConfig.useTabLayout
+                ? createTabBarLayout()
+                : createDrawerLayout();
+        };
 
-        // Show welcome modal only after layout is created for proper
-        // bookkeeping of the active view.
-        Promise.join(
-            appLayoutPromise,
-            welcomeModalControllerPromise,
-        function(menuController, welcomeModalController) {
-            // The welcome modal can be configured to show only
-            // once -- on initial startup, by passing in the
-            // parameter `{forced: false}` below
-            welcomeModalController.show({forced: true});
+        var initMainLayout = function() {
+            return createLayout().then(function(layoutController) {
+                // Deep linking services will enable deep linking on startup
+                // and while running it will open the deep link in the current
+                // active tab
+                deepLinkingServices = new DeepLinkingServices(layoutController);
+            });
+        };
 
-            // Deep linking services will enable deep linking on startup
-            // and while running
-            // It will open the deep link in the current active tab
-            deepLinkingServices = new DeepLinkingServices(menuController);
-        });
+        var runApp = function(previewedUrl) {
+            // TODO: [HYB-884] As a Scaffold developer,
+            // I would like for the baseURL to be set to the previewed URL
+
+            WelcomeModalController.init(errorControllerPromise)
+                .then(function(welcomeModalController) {
+                    // The welcome modal can be configured to show only once
+                    // (on first launch) by setting `{forced: false}` as the
+                    // parameter for welcomeModalController.show()
+                    welcomeModalController
+                        .show({forced: true})
+                        .finally(initMainLayout);
+                });
+        };
+
+        var runAppPreview = function() {
+            MobifyPreviewPlugin.init()
+                .then(function(previewPlugin) {
+                    previewPlugin
+                        .preview(BaseConfig.baseURL, BaseConfig.previewBundle)
+                        .then(runApp);
+                });
+        };
+
+        // Configure the previewEnabled flag located in baseConfig.js
+        // to enable/disable app preview
+        BaseConfig.previewEnabled ? runAppPreview() : runApp();
 
     }, undefined, true);
 };